Extract the cancel broadcast method from CommandDispatcher.
If we want to cancel requests from the broker, we'll want a method to let us do so.
We'll probably also want to synchronize on the broadcast and execution methods so that
we don't suffer from bad interleaving of requests to cancel and execute.  This is
still a living race condition, but at least it's a correct race condition.