Become less dependent on the execution map
It is possible that the command objects are being mutated
underneath our execution.  If that is happening, it will
cause three issues:
 * Null pointer exceptions when map.remove is run.
 * Command futures will never complete.
 * The execution map will fill unboundedly.

In order to fix this, insure that the resultFuture is
completed in every case.  In addition, save the identity
of the command that initiated the call, and if removal fails,
indicating that the command has been mutated and its hash has
altered, scan the map for a command with the same identity
and remove it.