Establish and use a path to throw unhandleable exceptions
We currently rely on the ability of CommandDispatcher to handle
catching RuntimeExceptions, but all such exceptions fall into the
UNKNOWN error category.  Here, we add a RuntimeException-derived
Exception that can carry an error code *a la* BaseException.  We
use this in one place where the error message resulting from the
exception singature is misleading, but we know what the error should
be.  Throwing it back to the top level will allow the error to be
properly categorized without unneccesarily cluttering the code with
checked exeception declarations.