Fixing Redirect Uri for Authenticator App, when migrating to MSAL
Address the following issue :
Broker redirect uri is invalid. Expected: urn:ietf:wg:oauth:2.0:oob Actual: msauth://.......

A test build was provided to the Authenticator app with this change and it all looks good,
we got a confirmation from the Authenticator team.

Initial PR which did not get into the prod to minimize the delta in the releases : https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/1399