Use android.net.NetworkCapabilities to check internet connectivity (#1887)
### What
Use android.net.NetworkCapabilities to check internet connectivity
ADO Item:
https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2017766/

### Why

[NetworkInfo](https://developer.android.com/reference/android/net/NetworkInfo)
class is marked as deprecated. It is recommended that callers should
instead use the
[ConnectivityManager.NetworkCallback](https://developer.android.com/reference/android/net/ConnectivityManager.NetworkCallback)
API to learn about connectivity changes, or switch to use
[ConnectivityManager#getNetworkCapabilities](https://developer.android.com/reference/android/net/ConnectivityManager#getNetworkCapabilities(android.net.Network))
or
[ConnectivityManager#getLinkProperties](https://developer.android.com/reference/android/net/ConnectivityManager#getLinkProperties(android.net.Network))
to get information synchronously.
Since we are checking the connectivity info synchronously before doing a
network operation switching to use getNetworkCapabilities for that
purpose.