Wrap Runnable with Current Telemetry Context (#1956)
This ensure that if spans or traces are started outside of a runnable,
then we can carry the same span or trace forward into the runnable. This
is particularly required for our partners that have integrated with our
broker open telemetry work and have also decided to create their own
spans wrapping around MSAL usage. In this case, we must carry the trace
information forward so we can create child spans with correct trace ids.

Prior Specs:

-
https://dev.azure.com/IdentityDivision/DevEx/_git/AuthLibrariesApiReview/pullrequest/6646?_a=files&path=/%5BAndroid%5D%20Telemetry/open_telemetry.md
-
https://dev.azure.com/IdentityDivision/DevEx/_git/AuthLibrariesApiReview/pullrequest/6841?_a=files&path=/%5BAndroid%5D%20Telemetry/span_context_propagation.md

Related:
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/1747