Create a custom safe Global Open Telemetry that doesn't crash (#1977)
A custom safe Open Telemetry Instance holder that doesn't crash on the
call to "get" Open Telemetry. The motivation behind this is that all the
Broker code has state that is per-process, whereas other libraries such
as Open Telemetry can carry state across processes. The default Global
Open Telemetry only allows initializing Open Telemetry once and
subsequent calls to initialize it will throw an error. This is
problematic for code such as ours that runs in multiple processes
because we end up initializing OTel twice (once in app process and once
in auth process).

For more details, see Javadoc