Adding in removal prompt dialog positive button string to string res files (#1979)
## Summary
Title summarizes it... since the string being added is the same phrase
as one added before ("Got it" in English), I C+P the strings from the
"smartcard_nfc_reminder_dialog_positive_button" string resource. This
string is used for the positive button in the smartcard removal prompt
dialog.
I'll be adding this string into my next localization PR to AuthApp for
consistency.