Adding Dialog Prompts to Remove Smartcard (at times that make sense) (#1966)
## Summary
### Why the CBA dialog flow is being changed again
Feedback from testers of the NFC dialog flow showed that it would be
beneficial to have dialog prompts that let the user know when they
should remove their smartcard. For example, after the first NFC tap,
once the client determines that there are/are not any errors and have
read off the certificates, we should let the user know that they do not
need to hold their smartcard to the device anymore.

There are two main places that a user should see such a dialog for NFC:
- After initially tapping in order to retrieve certs from the smartcard
(before the cert picker)
- After tapping in order to confirm PIN and sign with a private key

There is a potential use for the removal prompt dialog in the USB flow:
-After signing with the private key, and other processes involving the
smartcard have completed.
    - UPDATE: Upon positive feedback, this is included in this PR.

Including this prompt in the USB flow could potentially decrease the
amount of configuration change related errors, depending upon when most
users will plug in their smartcard. If they plug in their smartcard only
during the times our activity is running (when the authentication
WebView is showing), then implementing such a change will keep the
configuration changes occurring within our activities, which have the
appropriate configChanges flags. The flip side of this is that a user
who plugged in their smartcard within the host app activity will
experience a configuration change within the host app activity upon our
activities destroying. Logic has been added to detect if this is the
case, and no removal prompt will be displayed if a smartcard is still
detected by the end of the flow.

### What main changes were made
- Added a new `ShowSmartcardRemovalPromptDialog` method to
`DialogHolder` to be called when we want a user to remove their
smartcard (after NFC taps, before sending auth results, for NFC and
USB).
    - corresponding Dialog is `SmartcardRemovalPromptDialog`
- Made changes to `WebViewAuthorizationFragment` and
`AzureActiveDirectoryWebViewClient` to set a callback that sends the
auth result and finishes (in order to delay sending results until the
user removes their smartcard). This callback will be invoked immediately
if the cba challenge handler is null, and otherwise will be invoked
based on the status of the challenge handler. A callback is needed here
in order to make sure the user removes their smartcard before our
library's activity destroys.
- `onPausedSmartcardDiscovery` renamed to `prepForNextUserInteraction`.
Though we are stopping discovery in NFC's case, the emphasis is put on
giving some time to prepare for an expected user interaction, which
would be pausing for the user to remove their smartcard (in NFC case;
USB implementation of this method just invokes callback).
- <strike>Got rid of the `onCancelCba()` method in smartcard dialogs.
This was originally used to run logic when a smartcard was unexpectedly
unplugged. To keep more consistent with the `connectionCallback`
variable of the CBA managers, I created a `DisconnectedCallback` local
variable that should be set for unexpected/expected unplugging on the
USB side, or expected removal on the NFC side. </strike> Actually ended
up keeping the `onCancelCba()` logic, as I was overthinking things. The
method was renamed to be `onUnexpectedUnplug()` to be more specific and
to avoid confusion with the cancelCallbacks.
- Was able to get rid of callback generics for
`AbstractSmartcardCertBasedAuthManager`.
- <strike>Consolidated the local `cancelCallback` interfaces attached to
each dialog into a universal `ICancelCbaCallback`.</strike> Done in
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/1972#pullrequestreview-1316779182
- <strike>Consolidated the local `dismissCallback` interfaces attached
to each dialog into a universal `IDismissCallback`</strike> Done in
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/1975
-Added a `disconnect` method to the NFC manager that should disconnect a
smartcard and wait for the user to remove the smartcard.

### Testing
The Units tests associated with CBA have been updated to check for
`SmartcardRemovalPromptDialog` being shown at correct times.
I've manual tested different scenarios to make sure that the changes
made to `AzureActiveDirectoryWebVIewClient` does not interfere with
regular auth. I'll continue to manual test to make sure all bases that I
can think of are covered.

---------

Co-authored-by: Shahzaib <37125644+shahzaibj@users.noreply.github.com>