Adding YubiKit remove method back in; bumping YubiKit Versions (#1994)
## Summary
The remove() method in YubiKit allowed us to be able to detect when a
user would remove their YubiKey from an NFC connection. This method was
omitted from the previous CBA PR because there was some odd behavior
experienced with YubiKey removal for certain devices and not others
(removing the YubiKey wouldn't invoke the callback associated with
removal).
The YubiKit engineering team helped me investigate this month, and they
let me know that this was [a known issue occurring due to some changes
made in Android 13](https://github.com/Yubico/yubikit-android/pull/80).
A fix was made on their side in the latest YubiKit version, 2.2.0.

This PR re-adds the YubiKit remove() method call via our disconnect()
method and bumps the YubiKit Android and Piv versions. The dismissal
button was not removed from the removal prompt dialog in case users
would somehow get stuck in the flow again with a future Android OS
update.

## Testing
I extensively tested the NFC behavior on my Samsung testing device
(which worked before, but made sure no behavior broke), and I was able
to confirm that the behavior on an android device running on Android 13
was fixed as well.

## Version PRs
- https://github.com/AzureAD/ad-accounts-for-android/pull/2209
-
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/1784
- https://github.com/AzureAD/android-complete/pull/195