Allow MsalCpp to specify an InMemoryCache implementation of SharedPreferencesAccountCredentialCache (#1998)
This is a reimplementation of #1929, where the calling app needs to
choose to configure its token cache for this new in-memory caching
behavior. The goal here is that MSAL.CPP will specify this for now, and
other usages (broker, MSAL) could be brought over in the future in a
slower or optional rollout.

After consulting with Amit, and because the implementations here are so
different, it wasn't possible to reasonably maintain a single class with
forked implementation, and so instead the class itself has been forked.
Similarly the tests have been forked. I think it would be possible to
instead turn those into parameterized tests, and pull some of the setup
implementation as a preable to all test methods that uses that parameter
to choose whether to run against the memory-backed version or not. I
think that's still a fairly messy change, but I'm open to that option if
reviewers prefer it (or have a better idea).

---------

Co-authored-by: Nick Bopp <nichbop@microsoft.com>