Separate constants for max client and broker protocol versions (#2008)
The changes here are needed to enable flight for MSA accounts in Broker.
See related PR:
https://github.com/AzureAD/ad-accounts-for-android/pull/2235

The following changes have been made:

- Rename the constant for holding latest msal-to-broker protocol version
from `MSAL_TO_BROKER_PROTOCOL_VERSION_CODE` to
`LATEST_MSAL_TO_BROKER_PROTOCOL_VERSION_CODE`
- Bump the `LATEST_MSAL_TO_BROKER_PROTOCOL_VERSION_CODE` to `13.0` as
this is the version that supports MSA accounts in Broker.
- Add a new constant for MAX Protocol Version known by clients such as
MSAL. This is the same as LATEST MSAL BROKER protocol version (and
references that constant) but having its own constant makes it cleaner
to know what's being used in MSAL as parameter to CLIENT MAX.
- Add a new constant for MAX protocol Version known by Broker. This is
currently being to `12.0` because in the Broker we want the max to be
`12.0` as we will control this with a flight. For now we want MSA flight
disabled and therefore the default max is 12.0. The actual max used by
Broker is either this or 13.0 based on flight. See the tagged Broker PR
for more details.
- Also added a new constructor to BrokerMsalController for test
purposes. See broker PR for details.