Fix NPE in OTEL code (#2018)
Add a safe implementation of {@link ImplicitContextKeyed#makeCurrent()}
that doesn't crash. The default implementation in Open Telemetry
sometimes throws an NPE deep into OTel's code. Per our telemetry this is
happening on 0.08% of devices i.e. the impact is minimal. That said,
we're creating this safe wrapper here in effort to mitigate even that
minimal impact. If we get an NPE then we just return a Noop Scope.