Moving ClearCertPref call to clean up method (#2025)
### Summary
The C++ team mentioned that they're still seeing some
ERR_CERT_DATABASE_CHANGED errors pop up in a few apps, which we
attributed last time to the
[WebView.ClearClientCertPreferences](https://developer.android.com/reference/android/webkit/WebView#clearClientCertPreferences(java.lang.Runnable))
call that we have to make to allow CBA users to add more than one
account without having to restart the app.
This PR moves the one ClearClientCertPreferences call we have to the
cleanUp method of the OnDevice challenge handler. The good thing about
this move is that it will only get called if on-device CBA was ever
initiated, instead of every time the WebView gets initialized. (This
option has more recently been easily accessible due to the multiple
rounds of refactoring that have been done to CBA in the previous few
months :) ).
If this change doesn't improve the hit rate of the error the OneAuth
folks are seeing, I'll talk to the CBA team regarding the customer
impact of removing/keeping this method call.

### Testing
Tested logging in with CBA and making sure the
clearClientCertPreferences call isn't causing any abnormal behavior. Was
able to add another account without closing the app, and I was also able
to cancel out of the CBA flow, but then successfully add an account upon
restart of the WebView. Made sure that the method call wasn't being made
in a regular password login flow.
Also asked the C++ team if they can run this branch through their
automated tests to catch any issues on their side early.