Moving ClearClientCertPreferences to Factory instance (#2035)
### Summary
A change I made before to move ClearClientCertPreferences helped reduce
the amount of this method call in the on-device CBA instance, but it
introduced some regressions in user experience for smartcard CBA, as the
cert preferences weren't getting cleared, which would cause CBA to
always fail upon subsequent attempts to try to login (until they restart
the entire app, and broker app as well).
This PR puts the ClearClientCertPreferences method in the
CertBasedAuthFactory, where a boolean helps determine if CBA was
initiated, and clears if true.

### Testing
To be extra thorough:
- tested with mfatest@vimrang1, fidlab@msidlab4, and fidlab@msidlab3. 
- tested logging in with smartcard CBA, then on-device CBA.
- tested cancelling at various stages, including the dialog prompt to
select a type of CBA, within smartcard cba, within on-device cba,
errors.
- tested just logging in with a password, and made sure the
ClearClientCertPreferences method wasn't being called there.
- one thing to note is that with fidlab@msidlab3, the device auth
invokes CBA automatically. the logic in this PR doesn't clear the cert
pref in this case. this is a known thing, and in the majority of CBA
scenarios we've seen so far, they won't be using this type of device
auth. But this is something to investigate more in the future as we get
more users.