Replacing SHA-1 used in broker validation with SHA-512 (#2019)
### Summary
This is the first of a few rounds of PRs that address the use of SHA-1
in our libraries. This set of PRs focuses on replacing SHA-1 usage in
the broker verification logic with SHA-512.

Main changes in Common:
- The AuthenticationConstants file containing hardcoded signature hashes
(note: "signature" here refers to the [Android Signature
Class](https://developer.android.com/reference/android/content/pm/Signature),
which is an object representation of the signing certificate of an
app... kinda confusing) now also contains SHA-512 hashes for brokerHost,
Company Portal, and Authenticator. The BrokerData file has been updated
with the SHA-512 static strings accordingly.
- Teams IP Phone signature hash will be updated in a later PR. Its
current use shouldn't be affected by the changes (its used with the
original getCurrentSignatureForPackage method)
- The verifySignatureHash method now uses SHA-512 instead of SHA-1.
- The getCurrentSignatureForPackage static method now has a boolean
parameter (which determines whether to use SHA-512 or not). This method
is still used in many redirect URI and application identifier cases
(which will be changes that will occur in the future), so two methods
were created (getSha1SignatureForPackage and
getSha512SignatureForPackage). The uses related to broker validation are
using getSha512SignatureForPackage while redirect uri use cases will use
getSha1SignatureForPackage. More uses cases will be ported over to the
512 method in the future.

### Testing
- General logging in and logging out with msalTestApp and
Authenticator/Company Portal/BrokerHost as broker.
- Getting SSO Token (via BrokerHost broker api) 
- Device Token (via BrokerHost broker api) 
- Device signout unit tests passing

### Related PRs
- Broker: https://github.com/AzureAD/ad-accounts-for-android/pull/2251
- ADAL:
https://github.com/AzureAD/azure-activedirectory-library-for-android/pull/1732
- MSAL:
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/1826