Add android 14 check for skipping strong box isolation in pop token (#2053)
Created the PR on Fadi's behalf

Issue description : When a device is newly Intune enrolled in A14, the
user is able to login to Defender, but unable to retrieve the pop access
token. As a fallback, Defender allows onboarding without pop token
(which is why you were able to onboard fully to Defender). However, in
absence of the correct token Defender backend does not acknowledge
heartbeat calls we send every few hours to maintain device compliance.
Heartbeat failure subsequently fails the signals being sent to Intune
backend marking the device non-compliant
(Device_Threat_Health_Not_Reported reported by Intune).


This is an issue especially when policies are setup to make the device
noncompliant if Defender is not setup (this is what Microsoft has setup
internally).

Bug link :
https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2574078

Fix : There is new IAR requirement introduced for Android 14 that is
failing for Strong box isolation to work as expected. Skipping strong
box until a fix for this is made. We are still investigating on what
could be the long term fix. But until then, we will go ahead with this
temporary fix.

Testing : Shared the new broker build with CP team and they tested that
defender is working as expected after taking this fix.

---------

Co-authored-by: Fadi Durah <fadidurah@microsoft.com>