Clear cache of access tokens with an old application identifier field (#2058)
### Summary
_Note: A few unit tests are expected to fail until another PR which
updates variables
(https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2059)
is merged into this one. I separated it out in order to improve the
readability of this PR._

When created from the broker, access token records have an Application
Identifier string field attached to them which consists of the calling
app's package name and signing certificate hash:
`<Package name>/<signing certificate hash>`

The base-64 signing certificate hash is currently generated using SHA-1.
As we're currently making the effort to migrate our uses of SHA-1 to
SHA-512, I created a PR for the broker to now attach a SHA-512 hash to
the broker parameter Signature field, replacing the SHA-1 hash, for both
interactively and silently acquiring a token.

### Why this PR is also needed
Fortunately, the user shouldn't have to login again upon updating their
broker app; there will be an access token cache miss, but the unaffected
refresh token records will be able to reacquire an access token with the
updated SHA-512 application identifier field.
However, the access token with a SHA-1 application identifier will
remain in the cache indefinitely, as it won't be explicitly removed if
the broker is only searching for app identifiers including 512 hashes.

This PR addresses this issue by removing access tokens with a SHA-1
application identifier upon instantiation of the
SharedPreferencesAccountCredentialCache object (and the corresponding
MemoryCache version). This ensures that the cache is cleaned of access
tokens with an "old" identifier before any getCredentials() method is
called.

This logic should (in theory) run only once upon a user updating to the
version which includes said logic. I added a unit test which mimics this
scenario.

Related PRs:
-
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2059
: Will be merged into this one.
- https://github.com/AzureAD/ad-accounts-for-android/pull/2317 : The
mentioned broker PR, which will be merged _after_ this one.