Avoid NPE when Content-Type header is missing (#2079)
Ingesting 13.0.1, I hit NPEs in MSAL.CPP tests. We have a PoP test that
involves some test endpoint, and that endpoint is not setting a
Content-Type header. It looks like there is new telemetry code in 13.X
that isn't accounting for that header being missing in the way it does
others, and is generating an NPE.

---------

Co-authored-by: Nick Bopp <nichbop@microsoft.com>