Adding null checks for CancelHelper (CBA) (#2105)
## Summary
The OneAuth team let me know about some crashes associated with CBA. A
NPE is coming from `CertBasedAuthFactory.onCancelHelper` on the line
`mNfcSmartcardCertBasedAuthManager.clearConnectionCallback();`. The case
where `mNfcSmartcardCertBasedAuthManager` would be null is when a device
that is incapable of NFC initiates CBA. I missed putting a null check
here for the NFC and USB managers, as well in one other method (though
that one was protected by a null check before it's called, so it didn't
have a chance of causing issues in prod).
This PR adds in those null checks, as well as a unit test focusing on
null managers.