Add key derivation method that takes SecretKey object as argument (#2113)
### What
Add key derivation method that takes the key to derive from as SecretKey
object instead of Raw bytes.

### Why
Android KeyStore does not allow to extract raw key data. So, for the
KeyStore based key we will be using the SecretKey object to pass to the
Keyderivation function.

### How
Added a new method in SP800108KeyGen.java class that takes SecretKey
object as input and performs the keyderivation operation.

### Testing
Verified locally with a KeyStore based key
Also added a test to verify the derived key with default crypto factory.