Improve logging in BrokerMsalController.verifyBrokerVersionIsSupported (#2132)
Result bundle returned from Broker may or may not contain BrokerResult. 

After call returns from broker in BrokerMsalController, we attempt to
check if the error is unsupported_broker_version. To do so we try to
look for BrokerResult in the returned bundle, as the above error would
be returned as BrokerResult object.

But if it's not unsupported broker version error, then the returned
bundle may or may not contain BrokerResult object (depending on which
broker operation is executed). eg GetAccounts.. if BrokerResult is not
returned parsing result bundle raises exception (brokerResultFromBundle
method in MsalBrokerResultAdapter).

This exception is getting logged in verifyBrokerVersionIsSupported and
making the logs noisy with exception stack, even though the request
continues and is successful/failure based the code logic that follows.

Fix:
Updating the log statement to not log exception. We still log a message.