Create auth FIDO challenge object and factory (#2131)
## Summary
This PR adds new classes that represent FIDO challenges (auth) and
expected fields from the passkey server protocol. A factory is also
added that will return an IFidoChallenge based on the redirect URI
received. This factory will eventually include logic to return either an
AuthFidoChallenge or a RegFidoChallenge based on the design of the
passkey server protocol.
All of the query parameters of the auth passkey protocol except
allowedCredentials and keyTypes are required, so those two parameters
are nullable.

Note that these classes are not currently used or called anywhere, but
they will eventually be integrated in future PRs.
Unit tests were added for FidoChallengeFactory to test well-formed and
malformed redirect URIs.