Merged PR 9387: Use BuildConfig to set useRealAuthority and DC parameters
Sets `useRealAuthority` and `DC` parameters through the BuildConfig via gradle.

Note: this will set `useRealAuthority` to `true` by default now. The reasoning being that the 'default' behaviour should be to validate the incoming authority and to hit the authority passed in, and disabling this should only be used for development and testing.

To set either the `useRealAuthority` or `DC` parameters, set those variables in either the project or global `gradle.properties`. I'd recommend using the global properties file as that way it won't be impacted by switching branches and won't be committed to source control. Setting the parameters would look like the following:

```
useRealAuthority=true
dc=ESTS-PUB-SCUS-LZ1-FD000-TEST1
```

Note: Numerous test cases rely on `useRealAuthority` to be false in order to use the mock API, so when running local tests, it should be set to false. This is also the case for the CI pipelines, hence the update to the `azure-pipelines.yml`

Related work items: #2629229