FidoConstants (#2151)
### Summary
This PR creates the `FidoConstants` class, which currently has two
constants related to the WebAuthn extra query parameter. This query
parameter will be passed into MSAL by host apps when they want to
declare that they are WebAuthn capable.
The constants in this class will be used by msalTestApp, which will have
a switch to add the query parameter.