Catch strongbox exception on Android 14 (#2158)
**Issue** : This is related to the fix we did in the PR
https://github.com/AzureAD/microsoft-authentication-library-common-for-android/pull/2053

The fix we did previously worked and Defender team confirmed the same.
They were unblocked on Andorid 14. But, I started seeing the same issue
again on Android 14 because the "Build.VERSION.RELEASE_OR_CODENAME" is
different now. This is happening only on some devices (I think..
because, Defender team is not seeing this issue).
Build.VERSION.RELEASE_OR_CODENAME was previously UpsideDownCake. Now it
is coming as "14".

**Fix** : 

- I have added a new condition to check for
Build.VERSION.RELEASE_OR_CODENAME as 14 as well.
- Build.VERSION.RELEASE_OR_CODENAME was only available on Android API
level 30, so added a check for this.
- Also added some null checks.