Merged PR 9743: [Android SDK] Sample App refresh token is failing because access token in cache is null and causing sample app to crash
Problem:
When refreshing an expired access token, the token is deleted from the cache then parsed for the scopes that the user signed in with. If the refresh token fails subsequent attempts to refresh token will fail due to a NullPointerException because the access token in the cache is null.
This will cause offline usage of the SDK to remove access tokens and then subsequent refresh token attempts will fails (either offline or online).

Solution:
Do not delete the access token in the cache when refreshing expired tokens. The tokens should be overwritten when refreshed.
Additionally throw ServiceException when the access token read from the cache is null

Related work items: #2683265