WebAuthnJsonUtil; giving default value for UserVerificationPolicy (#2172)
### Summary
There are three main parts to this PR:
- `WebAuthnJsonUtil` class: helps convert a `FidoAuthChallenge` to a
JSON string which will be used as an input to the CredMan API in the
future.
- New dependency and plugin,
[kotlinx.serialization](https://github.com/Kotlin/kotlinx.serialization):
This was a recommended dependency for converting Kotlin classes into
JSON. Gson is known to be slow, and Moshi has some extra dependencies
needed specifically for Kotlin that include reflections, which will most
likely increase size and decrease speed at runtime.
Kotlinx.serialization is created for kotlin and (from the .jar size) is
relatively lighter. The versions selected correspond to our use of
kotlin 1.7.
- UserVerificationPolicy default value of "required": the server side
let me know that even though UserVerificationPolicy has multiple
possible values by WebAuthn standards, we should always expect it to be
"required" for AAD and MSA, and thus, the server side will never send it
to us in the protocol. I do want to still have the option available to
set that property in case the server side changes their mind, or
"required" doesn't end up being the only option in the future. So I
added a helper method to set the property as "required" by default, but
if the query parameter is present in the url, set it to whatever value
that may be.
### Related PRs
- https://github.com/AzureAD/android-complete/pull/251
-
https://github.com/AzureAD/microsoft-authentication-library-for-android/pull/1899
- https://github.com/AzureAD/ad-accounts-for-android/pull/2532
-
https://github.com/AzureAD/azure-activedirectory-library-for-android/pull/1757